# Backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/customer": {
      "get": {
        "tags": [
          "Customer"
        ],
        "description": "Get List of Customers",
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Customer"
        ],
        "description": "Create or Update User",
        "parameters": [
          {
            "name": "customer",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "phoneNo": {
                  "type": "string",
                  "example": "string"
                },
                "address": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/parts": {
      "get": {
        "tags": [
          "Parts"
        ],
        "description": "Get Parts",
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Parts"
        ],
        "description": "Create or Update Parts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/machines": {
      "get": {
        "tags": [
          "Machines"
        ],
        "description": "Get Machines",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Machines"
        ],
        "description": "Create or Update Machines",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contracts": {
      "put": {
        "tags": [
          "Contracts"
        ],
        "description": "Create Service Contracts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Contracts"
        ],
        "description": "Get Contracts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/repair": {
      "put": {
        "tags": [
          "Repair Jobs"
        ],
        "description": "Create Repair Jobs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/invoice": {
      "get": {
        "tags": [
          "Invoice"
        ],
        "description": "Get Invoice From certain Custoemr",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}